+++ docker/CHANGELOG.md
@@ -1,3 +1,5 @@
+# 0.1.14 
+[rajeshj]   - Adding hostname parameter to docker container
 # 0.1.13
 [rajeshj]   - Adding Write,Read and Kill timeout values for docker containers
 # 0.1.12